+++ commanderbot_ext/ext/automod/actions/abc/remove_roles_from_target_base.py
@@ -18,5 +18,7 @@ class RemoveRolesFromTargetBase(AutomodActionBase):
     async def apply(self, event: AutomodEvent):
         if member := self.get_target(event):
             guild: Guild = member.guild
+            # TODO Warn about unresolved roles. #logging
             roles = [guild.get_role(role_id) for role_id in self.roles]
+            roles = [role for role in roles if role]
             await member.remove_roles(roles)