+++ commanderbot_ext/ext/automod/conditions/message_content_contains.py
@@ -1,3 +1,4 @@
+import unicodedata
 from dataclasses import dataclass
 from typing import Optional, Tuple, Type, TypeVar
 
@@ -11,26 +12,36 @@ from commanderbot_ext.lib import JsonObject
 ST = TypeVar("ST")
 
 
+DEFAULT_NORMALIZATION_FORM = "NFKD"
+
+
 @dataclass
 class MessageContentContains(AutomodConditionBase):
     """
-    Check if message content contains a number of strings.
+    Check if message content contains a number of substrings.
 
     Attributes
     ----------
     contains
-        The strings to find.
+        The substrings to find. Unless `count` is specified, all substrings must be
+        found in order to pass.
     count
         The number of unique substrings to find. For example: a value of 1 requires any
         of the substrings to be found, whereas a value of 2 requires at least 2 to be
         found. If unspecified, all substrings must be found.
     ignore_case
         Whether to ignore upper vs lower case.
+    use_normalization
+        Whether to use unicode normalization or process the string as-is.
+    normalization_form
+        If enabled, the type of normalization to apply. Defaults to NFKD.
     """
 
     contains: Tuple[str]
     count: Optional[int] = None
     ignore_case: Optional[bool] = None
+    use_normalization: Optional[bool] = None
+    normalization_form: Optional[str] = None
 
     @classmethod
     def from_data(cls: Type[ST], data: JsonObject) -> ST:
@@ -47,6 +58,8 @@ class MessageContentContains(AutomodConditionBase):
             contains=contains,
             count=data.get("count"),
             ignore_case=ignore_case,
+            use_normalization=data.get("use_normalization"),
+            normalization_form=data.get("normalization_form"),
         )
 
     async def check(self, event: AutomodEvent) -> bool:
@@ -54,10 +67,16 @@ class MessageContentContains(AutomodConditionBase):
         # Short-circuit if there's no message or the message is empty.
         if not (message and message.content):
             return False
-        # Otherwise, check for a sufficient number of matches.
+        # Grab the message content.
         content = str(message.content)
+        # Normalize the message content, if enabled.
+        if self.use_normalization:
+            normalization_form = self.normalization_form or DEFAULT_NORMALIZATION_FORM
+            content = unicodedata.normalize(normalization_form, content)
+        # Convert the message content to lower-case, if we're ignoring case.
         if self.ignore_case:
             content = content.lower()
+        # Check for a sufficient number of substrings.
         remainder = self.count or len(self.contains)
         for substring in self.contains:
             # If the substring is found, adjust the counter.